[추가] 게임 횟수에 대한 예외 추가
게임 횟수가 0 이하일때 예외처리 구현

+++ src/main/java/logic/Player.java
@@ -16,7 +16,7 @@ public class Player {
     }
 
     private void init() {
-        gameCount = InputView.getGameCount();
+        gameCount = InputView.getGameCount().intValue();
         carList = InputView.getCarsList();
     }
